+++ src/Resources/app/storefront/src/main.js
@@ -1,5 +1,5 @@
-import CartFlyoutPlugin from './plugin/cart-flyout/cart-flyout.plugin';
+import CartFlyoutPlugin from "./plugin/cart-flyout/cart-flyout.plugin";
 
 const PluginManager = window.PluginManager;
 
-PluginManager.register('CartFlyoutPlugin', CartFlyoutPlugin, '.offcanvas-cart');
+PluginManager.register("CartFlyoutPlugin", CartFlyoutPlugin, ".offcanvas-cart");