+++ CHANGELOG.md
@@ -47,3 +47,16 @@
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/snippet/en-GB.json`
 -   `custom/plugins/StrixNLUxUpgrades/src/Resources/snippet/nl-NL.json`
 -   `custom/plugins/StrixNLUxUpgrades/src/Storefront/Page/Checkout/UnifiedDiscountSummarySubscriber.php`
+### 2025-09-26 15:30 — 0a866bd1
+- Added discount in summary to checkout/register
+
+**Changed files:**
+
+- `custom/plugins/StrixNLUxUpgrades/README.md`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/app/storefront/src/main.js`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/config/config.xml`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/config/services.xml`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/layout/footer/footer-language-widget.html.twig`
+- `custom/plugins/StrixNLUxUpgrades/src/Resources/views/storefront/layout/footer/footer.html.twig`
+- `custom/plugins/StrixNLUxUpgrades/src/Storefront/Page/FooterLanguageDataSubscriber.php`
+